feat: add more fields to pearl bank metrics
